+++ src/include/index/flat_pq_index.h
@@ -77,11 +77,11 @@ class flat_pq_index {
   // @todo Temporarily public to facilitate early testing
  public:
   // metadata
-  uint32_t dimensions_{0};
-  uint32_t num_subspaces_{0};
+  uint64_t dimensions_{0};
+  uint64_t num_subspaces_{0};
   size_t sub_dimensions_{0};
   size_t bits_per_subspace_{8};
-  uint32_t num_clusters_{256};
+  uint64_t num_clusters_{256};
   float tol_ = 0.001;
   size_t max_iter_ = 16;
   size_t num_threads_ = std::thread::hardware_concurrency();
@@ -114,10 +114,10 @@ class flat_pq_index {
    * @todo We don't really need dimension as an argument for any of our indexes
    */
   flat_pq_index(
-      uint32_t dimensions,
-      uint32_t num_subspaces,
+      uint64_t dimensions,
+      uint64_t num_subspaces,
       size_t bits_per_subspace = 8,
-      uint32_t num_clusters = 256)
+      uint64_t num_clusters = 256)
       : dimensions_(dimensions)
       , num_subspaces_(num_subspaces)
       , bits_per_subspace_(bits_per_subspace)
