+++ src/include/index/ivf_pq_index.h
@@ -196,7 +196,7 @@ class ivf_pq_index {
    ****************************************************************************/
 
   // Cached information about the partitioned vectors in the index
-  uint32_t dimensions_{0};
+  uint64_t dimensions_{0};
   uint64_t num_partitions_{0};
 
   // Cached information about the pq encoding
@@ -290,7 +290,7 @@ class ivf_pq_index {
    */
   ivf_pq_index(
       size_t nlist = 0,
-      uint32_t num_subspaces = 16,
+      uint64_t num_subspaces = 16,
       size_t max_iterations = 2,
       float convergence_tolerance = 0.000025f,
       float reassign_ratio = 0.075f,
@@ -1335,7 +1335,7 @@ class ivf_pq_index {
     return num_partitions_;
   }
 
-  uint32_t num_subspaces() const {
+  uint64_t num_subspaces() const {
     return num_subspaces_;
   }
