+++ src/include/index/vamana_index.h
@@ -135,7 +135,7 @@ class vamana_index {
    ****************************************************************************/
 
   // Cached information about the index
-  uint32_t dimensions_{0};
+  uint64_t dimensions_{0};
   uint64_t num_vectors_{0};
   uint64_t num_edges_{0};
 
@@ -657,7 +657,7 @@ class vamana_index {
     return std::make_tuple(std::move(top_k_scores), std::move(top_k));
   }
 
-  constexpr uint32_t dimensions() const {
+  constexpr uint64_t dimensions() const {
     return dimensions_;
   }
